chat example updated
